split the Halos_lens
some debug
